BI_trabalho/requirements.txt
streamlit==1.31.0
opencv-python-headless==4.9.0.80
pyzbar==0.1.9
pandas==2.2.0
Pillow==10.2.0
numpy==1.26.3
